refactor getBase & getSolids
