--- examples/src/meshing/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" href="/img/favicon.ico" type="image/x-icon" />
-    <title>Awatif Examples - Meshing</title>
-    <!-- Google tag (gtag.js) -->
-    <script
-      async
-      src="https://www.googletagmanager.com/gtag/js?id=G-1X49GCVEK9"
-    ></script>
-    <script>
-      window.dataLayer = window.dataLayer || [];
-      function gtag() {
-        dataLayer.push(arguments);
-      }
-      gtag("js", new Date());
-
-      gtag("config", "G-1X49GCVEK9");
-    </script>
-  </head>
-  <body>
-    <script type="module" src="./main.ts"></script>
-  </body>
-</html>
