+++ examples/src/solids/main.ts
@@ -1,24 +1,11 @@
 import van, { State } from "vanjs-core";
-import {
-  Mesh,
-  Shape,
-  ExtrudeGeometry,
-  MeshPhongMaterial,
-  BufferGeometry,
-  Path,
-  Vector2,
-  BufferAttribute,
-  Matrix4,
-  LineSegments,
-  LineBasicMaterial,
-  Object3D,
-  Float32BufferAttribute,
-} from "three";
-import { mergeGeometries } from "three/examples/jsm/utils/BufferGeometryUtils.js";
-import { Element, Mesh as FemMesh } from "awatif-fem";
+import { Object3D } from "three";
+import { Mesh as FemMesh } from "awatif-fem";
 import { Parameters, getParameters, getToolbar, getViewer } from "awatif-ui";
 import { Building, ColumnData, SlabData } from "./data-model";
 import { meshing } from "./meshing";
+import { base, getBaseGeometry } from "./getBase";
+import { getSolidsGeometry, solids as solidsMesh } from "./getSolids";
 
 type SimpleBuilding = {
   points: State<number[][]>;
@@ -85,16 +72,9 @@ const parameters: Parameters = {
   stories: { value: van.state(2), min: 1, max: 5, step: 1 },
 };
 
-const solidsMesh = new Mesh(
-  new BufferGeometry(),
-  new MeshPhongMaterial({ color: 0xffe6cc })
-);
-const base = new LineSegments(new BufferGeometry(), new LineBasicMaterial());
-base.frustumCulled = false;
-base.material.depthTest = false; // don't know why but is solves the rendering order issue
-
 const objects3D: State<Object3D[]> = van.state([base]);
 const solids: State<Object3D[]> = van.state([solidsMesh]);
+
 const mesh: FemMesh = {
   nodes: van.state([]),
   elements: van.state([]),
@@ -209,237 +189,10 @@ van.derive(() => {
 
 document.body.append(
   getParameters(parameters),
-  getViewer({ objects3D, solids, mesh }),
+  getViewer({ objects3D, solids, mesh, settingsObj: { loads: false } }),
   getToolbar({
     sourceCode:
       "https://github.com/madil4/awatif/blob/main/examples/src/solids/main.ts",
     author: "https://www.linkedin.com/in/abderrahmane-mazri-4638a81b8/",
   })
 );
-
-//Utils
-function getBaseGeometry(
-  points: number[][],
-  slabs: number[][][],
-  columns: number[][]
-): BufferGeometry {
-  const geometry = new BufferGeometry();
-
-  const columnsVertices = columns
-    .map((column) => [points[column[0]], points[column[1]]].flat())
-    .flat();
-
-  // borrowed from viewer elements object
-  const slabsVertices = slabs
-    .map((slab) =>
-      elementToEdges(slab[0])
-        .map((edge) => [...points[edge[0]], ...points[edge[1]]])
-        .flat()
-    )
-    .flat();
-
-  geometry.setAttribute(
-    "position",
-    new Float32BufferAttribute([...columnsVertices, ...slabsVertices], 3)
-  );
-
-  return geometry;
-
-  function elementToEdges(element: number[]): Element[] {
-    if (element.length === 2) return [element];
-
-    const edges: [number, number][] = [];
-
-    for (let i = 0; i < element.length; i++) {
-      edges.push([element[i], element[(i + 1) % element.length]]);
-    }
-
-    return edges;
-  }
-}
-
-function getSolidsGeometry(
-  points: number[][],
-  slabs: number[][][],
-  columns: number[][]
-): BufferGeometry {
-  const columnWidth: number = 0.3;
-  const columnHeight: number = 0.3;
-
-  const slabsGeometry = createSlabsGeometry(points, slabs);
-  const columnsGeometry = createColumnsGeometry(points, columns);
-
-  return mergeGeometries([slabsGeometry, columnsGeometry]);
-
-  function createSlabsGeometry(
-    points: number[][],
-    slabsIndices: number[][][],
-    slabHeight: number = 0.3
-  ): BufferGeometry {
-    const buffers: BufferGeometry[] = [];
-
-    for (let k = 0; k < slabsIndices.length; k++) {
-      for (let i = 0; i < slabsIndices[k].length; i++) {
-        const contour: number[][] = [];
-        for (let j = 0; j < slabsIndices[k][i].length; j++) {
-          const pointIdx = slabsIndices[k][i][j];
-          contour.push(points[pointIdx]);
-        }
-
-        const offsetedContour = offsetContour(contour, columnWidth / 2);
-
-        const slabShape = new Shape();
-        const hole = new Path();
-        for (let i = 0; i < offsetedContour.length; i++) {
-          if (i == 0)
-            slabShape.moveTo(offsetedContour[0][0], offsetedContour[0][1]);
-          else slabShape.lineTo(offsetedContour[i][0], offsetedContour[i][1]);
-        }
-
-        const geometry = new ExtrudeGeometry(slabShape, {
-          depth: slabHeight,
-          bevelEnabled: false,
-        });
-
-        geometry.translate(0, 0, offsetedContour[0][2] - slabHeight / 2);
-        buffers.push(geometry);
-      }
-    }
-
-    return mergeGeometries(buffers);
-
-    function offsetContour(
-      contour: number[][],
-      offset: number = 0
-    ): number[][] {
-      const result: number[][] = [];
-
-      const _contour: Vector2[] = [];
-
-      for (let i = 0; i < contour.length; i++)
-        _contour.push(new Vector2(contour[i][0], contour[i][1]));
-
-      let _offset = new BufferAttribute(new Float32Array([offset, 0, 0]), 3);
-
-      for (let i = 0; i < _contour.length - 1; i++) {
-        let v1 = new Vector2().subVectors(
-          _contour[i - 1 < 0 ? _contour.length - 1 : i - 1],
-          _contour[i]
-        );
-        let v2 = new Vector2().subVectors(
-          _contour[i + 1 == _contour.length ? 0 : i + 1],
-          _contour[i]
-        );
-        let angle = v2.angle() - v1.angle();
-        let halfAngle = angle * 0.5;
-
-        let hA = halfAngle;
-        let tA = v2.angle() + Math.PI * 0.5;
-
-        let shift = Math.tan(hA - Math.PI * 0.5);
-        let shiftMatrix = new Matrix4().set(
-          1,
-          0,
-          0,
-          0,
-          -shift,
-          1,
-          0,
-          0,
-          0,
-          0,
-          1,
-          0,
-          0,
-          0,
-          0,
-          1
-        );
-
-        let tempAngle = tA;
-        let rotationMatrix = new Matrix4().set(
-          Math.cos(tempAngle),
-          -Math.sin(tempAngle),
-          0,
-          0,
-          Math.sin(tempAngle),
-          Math.cos(tempAngle),
-          0,
-          0,
-          0,
-          0,
-          1,
-          0,
-          0,
-          0,
-          0,
-          1
-        );
-
-        let translationMatrix = new Matrix4().set(
-          1,
-          0,
-          0,
-          _contour[i].x,
-          0,
-          1,
-          0,
-          _contour[i].y,
-          0,
-          0,
-          1,
-          0,
-          0,
-          0,
-          0,
-          1
-        );
-
-        let cloneOffset = _offset.clone();
-        cloneOffset.needsUpdate = true;
-        cloneOffset.applyMatrix4(shiftMatrix);
-        cloneOffset.applyMatrix4(rotationMatrix);
-        cloneOffset.applyMatrix4(translationMatrix);
-
-        result.push([cloneOffset.getX(0), cloneOffset.getY(0), contour[i][2]]);
-      }
-
-      return result;
-    }
-  }
-
-  function createColumnsGeometry(
-    points: any[][],
-    columnsIndices: number[][]
-  ): BufferGeometry {
-    const buffers: BufferGeometry[] = [];
-    const columnShape = new Shape();
-
-    // columns base
-    columnShape.lineTo(0 + columnWidth, 0);
-    columnShape.lineTo(0 + columnWidth, 0 + columnHeight);
-    columnShape.lineTo(0, 0 + columnHeight);
-
-    for (let i = 0; i < columnsIndices.length; i++) {
-      const p1 = points[columnsIndices[i][0]];
-      const p2 = points[columnsIndices[i][1]];
-
-      const height = p2[2] - p1[2];
-
-      const geometry = new ExtrudeGeometry(columnShape, {
-        depth: height,
-        bevelEnabled: false,
-      });
-
-      geometry.translate(
-        p1[0] - columnWidth / 2,
-        p1[1] - columnHeight / 2,
-        p1[2]
-      );
-
-      buffers.push(geometry);
-    }
-
-    return mergeGeometries(buffers);
-  }
-}
